refactor(adk): rename media module and source Vertex project from GC_PROJECT_ID
Address maintainer review on PR #82:
- Rename gcs_signing.py -> media_filedata.py: the module no longer signs GCS
  URLs; it converts media references to gs:// URIs and injects them as FileData
  parts for Vertex. Update the import in agent.py accordingly.
- Use the established GC_PROJECT_ID secret for the Vertex AI project instead of
  introducing a separate GOOGLE_CLOUD_PROJECT secret. The container env var keeps
  its SDK-required name (GOOGLE_CLOUD_PROJECT) but is now fed from ${GC_PROJECT_ID}
  in service.template.yaml, and the redundant deploy.yml env/override is removed.
