+++ service.template.yaml
@@ -48,7 +48,7 @@ spec:
         - name: GOOGLE_GENAI_USE_VERTEXAI
           value: "true"
         - name: GOOGLE_CLOUD_PROJECT
-          value: "${GOOGLE_CLOUD_PROJECT}"
+          value: "${GC_PROJECT_ID}"
         - name: GOOGLE_CLOUD_LOCATION
           value: "global"
         - name: LANGFUSE_PUBLIC_KEY
